Show waiting status after pressing end turn
